Add a configurable handler for hosts file updates
When the hosts file is updated, it can be useful to send a notification
to some service like e.g. dnsmasq so that changes in the hosts file are
taken into use. Add a configurable handler name to all the tasks in this
role that update the hosts file. The default handler is undefined,
meaning by default no action is taken when the hosts file is updated.