Use include_tasks instead of include
It's new with ansible 2.4. The include: is still there in 2.5 but it's deprecated
and it looks like it's similar to import: - meaning it will skip a lot
of things.

With several hundreds of hosts in the inventory to skip running a
playbook that uses the templating style the playbook went from 3m6s
to 2m21s or a reduction of 45 seconds.

Also a note about using "| sort" jinja2 filter on the groups to get the
list of hosts sorted